move annoMatrix to src dir
